Move some things around to make it easier to grok
